refactor: move Loader under Configuration namespace
